requirements.txt
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
pdfplumber==0.10.3
python-docx==0.8.11
nltk==3.8.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
openpyxl==3.1.2
google-generativeai==0.3.1
mysql-connector-python==8.2.0
